# sha: reuse cached EVP_MD on one-shot functions