This is a continuation PR of #75, in which I missed to reuse the `EVP-MD` objects on one-shot functions.

The perf improvement is noticeable:

```cmd
goos: linux
goarch: amd64
pkg: github.com/golang-fips/openssl
cpu: AMD EPYC 7763 64-Core Processor                
          │   old.txt   │               new.txt               │
          │   sec/op    │   sec/op     vs base                │
SHA256-16   549.4n ± 1%   287.9n ± 1%  -47.59% (p=0.000 n=10)

          │   old.txt    │               new.txt                │
          │     B/s      │     B/s       vs base                │
SHA256-16   13.89Mi ± 1%   26.49Mi ± 1%  +90.80% (p=0.000 n=10)

          │  old.txt   │            new.txt             │
          │    B/op    │    B/op     vs base            │
SHA256-16   0.000 ± 0%   0.000 ± 0%  ~ (p=1.000 n=10) ¹
¹ all samples are equal

          │  old.txt   │            new.txt             │
          │ allocs/op  │ allocs/op   vs base            │
SHA256-16   0.000 ± 0%   0.000 ± 0%  ~ (p=1.000 n=10) ¹
¹ all samples are equal
```